test: split helpers_test.go into focused files
- helpers_validation_test.go: color/priority/type validation tests
- helpers_label_test.go: GetLabelByID tests
- helpers_column_test.go: column helper tests (FindColumnByName, FormatAvailableColumns, GetCurrentColumnName)
- helpers_project_test.go: GetProjectID tests
